+++ CHANGELOG.md
@@ -1,4 +1,12 @@
+
 # Changelog for rabtap
+
+
+## v1.16 (2019-03-15)
+
+* new option `--by-connection` for info command added, making `info` show
+  connection > 
+* new: prefetch count added to output of `info` command (on consumer level)
 
 ## v1.15 (2019-03-01)